Raised version numbers of Extbase and Fluid to 1.3.0RC1
git-svn-id: https://svn.typo3.org/TYPO3v4/Core/trunk@10208 709f56b5-9817-0410-a4d7-c38de5d9e867

+++ typo3/sysext/extbase/Classes/Persistence/Repository.php
@@ -144,7 +144,10 @@ class Tx_Extbase_Persistence_Repository implements Tx_Extbase_Persistence_Reposi
 		}
 
 		$this->addedObjects->attach($object);
-		$this->removedObjects->detach($object);
+
+		if ($this->removedObjects->contains($object)) {
+			$this->removedObjects->detach($object);
+		}
 	}
 
 	/**
@@ -161,7 +164,9 @@ class Tx_Extbase_Persistence_Repository implements Tx_Extbase_Persistence_Reposi
 
 		if ($this->addedObjects->contains($object)) {
 			$this->addedObjects->detach($object);
-		} else {
+		}
+
+		if (!$object->_isNew()) {
 			$this->removedObjects->attach($object);
 		}
 	}
